api/requirements.txt
Django==5.0.7
psycopg[binary,pool]==3.2.1
djangorestframework==3.15.2
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.1
django-filter==24.2
dj-database-url==2.2.0
python-dotenv==1.0.1
pandas==2.2.2
openpyxl==3.1.5
Gunicorn==22.0.0
